feat: add custom hook
